% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_events.R
\name{get_git_events}
\alias{get_git_events}
\title{Get a list of public events from GitHub}
\usage{
get_git_events()
}
\arguments{
\item{path}{string}
}
\value{
list
}
\description{
Returns the 30 most recent public events from Github (within the last 90 days), including all metadata, as a list of lists.
}
\examples{
get_git_events()
}
